[FIX] im_livechat: remove non-required dependencies
This commit reduces the size and avoid conflicts caused by
the embed live chat bundle: jquery, legacy libs...

At the same time, this commit fixes the emoji picker that was
wrongly positionned and scrolled up the page when opening.

Finally, this commit ensures the chat window is always at the
bottom of the screen.

opw-3539362

X-original-commit: 1b29419e50343bfd723b41368ba38ce63b063d92
