+++ addons/im_livechat/static/src/embed/external/overrides.scss
@@ -1,0 +1,3 @@
+.o-main-components-container, .o_notification_manager {
+    position: fixed !important;
+}
