+++ addons/mail/static/src/core/common/message.js
@@ -18,7 +18,6 @@ import {
     Component,
     onMounted,
     onPatched,
-    onWillUnmount,
     useChildSubEnv,
     useEffect,
     useRef,
@@ -99,8 +98,6 @@ export class Message extends Component {
             isHovered: false,
             isClicked: false,
             expandOptions: false,
-            lastReadMoreIndex: 0,
-            isReadMoreByIndex: new Map(),
             originalEmail: false,
             emailHeaderOpen: false,
         });
@@ -150,15 +147,14 @@ export class Message extends Component {
         });
         onMounted(() => {
             if (this.messageBody.el) {
-                $(this.messageBody.el).find(".o-mail-read-more-less").remove();
-                this.insertReadMoreLess($(this.messageBody.el));
+                this.prepareMessageBody(this.messageBody.el);
             }
             if (this.shadowBody.el) {
                 this.shadowRoot = this.shadowBody.el.attachShadow({ mode: "open" });
                 const body = document.createElement("span");
                 body.innerHTML =
                     this.props.messageSearch?.highlight(this.message.body) ?? this.message.body;
-                this.insertReadMoreLess($(body));
+                this.prepareMessageBody(body);
                 const color = cookie.get("color_scheme") === "dark" ? "white" : "black";
                 this.shadowStyle = document.createElement("style");
                 this.shadowStyle.innerHTML = `
@@ -178,11 +174,6 @@ export class Message extends Component {
                 `;
                 this.shadowRoot.appendChild(this.shadowStyle);
                 this.shadowRoot.appendChild(body);
-            }
-        });
-        onWillUnmount(() => {
-            if (this.messageBody.el) {
-                $(this.messageBody.el).find(".o-mail-read-more-less").remove();
             }
         });
     }
@@ -436,6 +427,11 @@ export class Message extends Component {
         this.enterEditMode();
     }
 
+    /**
+     * @param {HTMLElement} element
+     */
+    prepareMessageBody(element) {}
+
     enterEditMode() {
         const messageContent = convertBrToLineBreak(this.props.message.body);
         this.props.message.composer = {
@@ -471,91 +467,6 @@ export class Message extends Component {
                 mail_message_to_resend: this.message.id,
             },
         });
-    }
-
-    /**
-     * Modifies the message to add the 'read more/read less' functionality
-     * All element nodes with 'data-o-mail-quote' attribute are concerned.
-     * All text nodes after a ``#stopSpelling`` element are concerned.
-     * Those text nodes need to be wrapped in a span (toggle functionality).
-     * All consecutive elements are joined in one 'read more/read less'.
-     *
-     * FIXME This method should be rewritten (task-2308951)
-     *
-     * @param {jQuery} $element
-     */
-    insertReadMoreLess($element) {
-        const groups = [];
-        let readMoreNodes;
-
-        // nodeType 1: element_node
-        // nodeType 3: text_node
-        const $children = $element
-            .contents()
-            .filter(
-                (index, content) =>
-                    content.nodeType === 1 || (content.nodeType === 3 && content.nodeValue.trim())
-            );
-
-        for (const child of $children) {
-            let $child = $(child);
-
-            // Hide Text nodes if "stopSpelling"
-            if (child.nodeType === 3 && $child.prevAll('[id*="stopSpelling"]').length > 0) {
-                // Convert Text nodes to Element nodes
-                $child = $("<span>", {
-                    text: child.textContent,
-                    "data-o-mail-quote": "1",
-                });
-                child.parentNode.replaceChild($child[0], child);
-            }
-
-            // Create array for each 'read more' with nodes to toggle
-            if (
-                $child.attr("data-o-mail-quote") ||
-                ($child.get(0).nodeName === "BR" &&
-                    $child.prev('[data-o-mail-quote="1"]').length > 0)
-            ) {
-                if (!readMoreNodes) {
-                    readMoreNodes = [];
-                    groups.push(readMoreNodes);
-                }
-                $child.hide();
-                readMoreNodes.push($child);
-            } else {
-                readMoreNodes = undefined;
-                this.insertReadMoreLess($child);
-            }
-        }
-
-        for (const group of groups) {
-            const index = this.state.lastReadMoreIndex++;
-            // Insert link just before the first node
-            const $readMoreLess = $("<a>", {
-                class: "o-mail-read-more-less d-block",
-                href: "#",
-                text: "Read More",
-            }).insertBefore(group[0]);
-
-            // Toggle All next nodes
-            if (!this.state.isReadMoreByIndex.has(index)) {
-                this.state.isReadMoreByIndex.set(index, true);
-            }
-            const updateFromState = () => {
-                const isReadMore = this.state.isReadMoreByIndex.get(index);
-                for (const $child of group) {
-                    $child.hide();
-                    $child.toggle(!isReadMore);
-                }
-                $readMoreLess.text(isReadMore ? "Read More" : "Read Less");
-            };
-            $readMoreLess.click((e) => {
-                e.preventDefault();
-                this.state.isReadMoreByIndex.set(index, !this.state.isReadMoreByIndex.get(index));
-                updateFromState();
-            });
-            updateFromState();
-        }
     }
 
     openReactionMenu() {
